+++ Packages/com.unity.inputsystem/InputSystem/Actions/InputBinding.cs
@@ -425,6 +425,9 @@ namespace UnityEngine.InputSystem
 
         [SerializeField] private string m_Name;
         [SerializeField] internal string m_Id;
+        [Tooltip("Path of the control to bind to. Matched at runtime to controls from InputDevices present at the time.\n\nCan either be "
+            + "graphically from the control picker dropdown UI or edited manually in text mode by clicking the 'T' button. Internally, both "
+            + "method result in control path strings that look like, for example, \"<Gamepad>/buttonSouth\".")]
         [SerializeField] private string m_Path;
         [SerializeField] private string m_Interactions;
         [SerializeField] private string m_Processors;